+++ packages/next-common/context/post/democracy/referendum/result.js
@@ -1,0 +1,7 @@
+import { useTimelineData } from "../../index";
+
+export default function useIsDemocracyPassed() {
+  const timeline = useTimelineData();
+  const item = timeline.reverse().find(item => "Passed" === item.method);
+  return !!item;
+}