+++ packages/next/components/referenda/vote.js
@@ -1,28 +1,22 @@
 import { memo, useCallback, useState } from "react";
 import dynamic from "next/dynamic";
 import styled from "styled-components";
-import { calcPassing } from "utils/referendumUtil";
 import useApi from "next-common/utils/hooks/useApi";
 import Loading from "next-common/components/loading";
-import Chains from "next-common/utils/consts/chains";
 import { SecondaryCardDetail } from "next-common/components/styled/containers/secondaryCard";
 import { TitleContainer } from "next-common/components/styled/containers/titleContainer";
 import { useDispatch, useSelector } from "react-redux";
-import { latestHeightSelector } from "next-common/store/reducers/chainSlice";
 import {
-  electorateSelector,
   fetchReferendumStatus,
   isLoadingElectorateSelector,
   isLoadingReferendumStatusSelector,
   isLoadingVotesSelector,
-  referendumStatusSelector,
   votesSelector,
 } from "next-common/store/reducers/referendumSlice";
 import SubLink from "next-common/components/styled/subLink";
 import VoteBar from "./voteBar";
 import TallyInfo from "next-common/components/referenda/tally/info";
 import { emptyFunction } from "next-common/utils";
-import { useChain } from "next-common/context/chain";
 import MyVote from "./myVote";
 import SecondaryButton from "next-common/components/buttons/secondaryButton";
 import {
@@ -51,6 +45,11 @@ const VotesGroupItems = styled.div`
   ${items_center};
   ${gap_x(12)};
 `;
+import useDemocracyTally from "next-common/context/post/democracy/referendum/tally";
+import useIsDemocracyPassing from "next-common/context/post/democracy/referendum/passing";
+import useIsDemocracyVoteFinished from "next-common/context/post/democracy/referendum/isVoteFinished";
+import useDemocracyThreshold from "next-common/context/post/democracy/referendum/threshold";
+import useIsDemocracyPassed from "next-common/context/post/democracy/referendum/result";
 
 const VotePopup = dynamic(() => import("components/referenda/popup"), {
   ssr: false,
@@ -105,28 +104,25 @@ const RejectStatus = styled(Status)`
   background: ${(props) => props.theme.secondaryRed100};
 `;
 
-function Vote({
-  referendumInfo,
-  referendumIndex,
-  onFinalized = emptyFunction,
-}) {
-  const chain = useChain();
+function Vote({ referendumIndex, onFinalized = emptyFunction }) {
   const dispatch = useDispatch();
   const [showVote, setShowVote] = useState(false);
   const [showFlattenedVotesList, setShowFlattenedVotesList] = useState(false);
   const [showNestedVotesList, setShowNestedVotesList] = useState(false);
   const [showCallsVotesList, setShowCallsVotesList] = useState(false);
   const api = useApi();
-  const blockHeight = useSelector(latestHeightSelector);
+  const tally = useDemocracyTally();
+  const isPassing = useIsDemocracyPassing();
+  const threshold = useDemocracyThreshold();
+  const isPassed = useIsDemocracyPassed();
 
-  const electorate = useSelector(electorateSelector);
   const isElectorateLoading = useSelector(isLoadingElectorateSelector);
   const isLoadingVotes = useSelector(isLoadingVotesSelector);
   const { allAye = [], allNay = [] } = useSelector(votesSelector);
-  const referendumStatus = useSelector(referendumStatusSelector);
   const isLoadingReferendumStatus = useSelector(
     isLoadingReferendumStatusSelector,
   );
+  const isVoteFinished = useIsDemocracyVoteFinished();
 
   const [updateTime, setUpdateTime] = useState(0);
 
@@ -135,23 +131,13 @@ function Vote({
     setUpdateTime(Date.now());
   }, [dispatch, api, referendumIndex]);
 
-  const isPassing = calcPassing(referendumStatus, electorate);
-
-  const isKsm = Chains.kusama === chain;
-  const finished =
-    referendumInfo?.finished ||
-    (isKsm && referendumIndex < 198) ||
-    blockHeight > referendumStatus?.end;
-
   let finishedResult;
-  if (referendumInfo?.finished) {
-    finishedResult = referendumInfo?.finished?.approved ? (
+  if (isVoteFinished) {
+    finishedResult = isPassed ? (
       <PassStatus>Passed</PassStatus>
     ) : (
       <RejectStatus>Failed</RejectStatus>
     );
-  } else if (isKsm && referendumIndex < 5) {
-    finishedResult = <PassStatus>Passed</PassStatus>;
   }
 
   return (
@@ -167,22 +153,19 @@ function Vote({
         </Title>
 
         <VoteBar
-          tally={referendumStatus?.tally}
-          electorate={electorate}
-          threshold={referendumStatus?.threshold}
+          tally={tally}
+          electorate={tally.electorate}
+          threshold={threshold}
         />
 
         <TallyInfo
-          tally={referendumStatus?.tally}
-          electorate={electorate}
           isLoadingVotes={isLoadingVotes}
           allAye={allAye}
           allNay={allNay}
         />
 
         {finishedResult}
-        {referendumInfo &&
-          !finished &&
+        {!isVoteFinished &&
           (isPassing ? (
             <PassStatus>Passing</PassStatus>
           ) : (
@@ -206,7 +189,7 @@ function Vote({
         <MyVote updateTime={updateTime} />
       </SecondaryCardDetail>
 
-      {!finished && (
+      {!isVoteFinished && (
         <SecondaryButton
           onClick={() => {
             setShowVote(true);